requirements.txt
Pillow==6.1.0
numpy==1.16.6
opencv-python-headless==4.2.0.32
colormath==3.0.0
pytesseract==0.3.0

aggdraw==1.3.11
colour==0.1.5
geographiclib==1.50
pyproj==2.2.2
pyshp==2.1.0